upgrade flow-bin from 0.95.1 to 0.105.2
